+++ users/drashna/rgb/rgb_stuff.c
